HW6 Upload
I created a simple start menu by following a tutorial online and after getting it working and switching to the main game, I compressed the unity package and it was still too large for GitHub. So I went in and deleted some assets that I wasn't using to get it to size. After checking everything seemed to work except the spawn of the enemies and I couldn't figure out if it was an issue with the sprite or the prefab.